Dedupe the domain stuff
=============================== Coverage summary ================
Statements   : 89.23% ( 11356/12727 ), 12 ignored
Branches     : 82.93% ( 5030/6065 ), 71 ignored
Functions    : 90.84% ( 843/928 ), 11 ignored
Lines        : 89.78% ( 10083/11231 )
=================================================================
