+++ src/pre/dsl2ml.js
@@ -49,7 +49,7 @@ import {
 } from './ml';
 import {
   domain_createRange,
-} from './domain';
+} from '../domain';
 
 // BODY_START
